app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.3
django-ckeditor==6.7.3
django-debug-toolbar==6.1.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-property-filter==1.3.0
django-view-breadcrumbs==2.5.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
